Port to MSVC. Work in progress; only oblcore builds now.
Changed regexp implementation for MSVC from glibc regexp to pcre2. The glibc source
does not compile with MSVC.

+++ include/data-typedefs.h
@@ -165,7 +165,7 @@ typedef struct _typedescr {
   int       promote_to;
   int      *ancestors;
   int      *implements;
-  int       implements_sz;
+  size_t    implements_sz;
   int       count;
   int      *inherits;
 } typedescr_t;
